requirements.txt
selenium
googletrans==4.0.0rc1
fastapi
uvicorn
playwright
sentence-transformers
chromadb
requests
pillow
pytesseract